requirements-dev.txt
# Development and build dependencies
# Includes packaging tools and development utilities

# ============================================
# Packaging tools
# ============================================
pyinstaller>=6.0.0

# ============================================
# Configuration file handling
# ============================================
pyyaml>=6.0

# ============================================
# Progress visualization
# ============================================
rich>=13.0.0

# ============================================
# Testing tools
# ============================================
pytest>=7.0.0
pytest-cov>=4.0.0
